@@ -6,20 +6,37 @@ require_once __DIR__ . '/../vendor/autoload.php';
 
 use PhpAnonymizer\Anonymizer\AnonymizerBuilder;
 use PhpAnonymizer\Anonymizer\Enum\NodeParser;
+use PhpAnonymizer\Anonymizer\Enum\RuleSetParser;
 
 $anonymizer = (new AnonymizerBuilder())
     ->withDefaults()
-    // set nodeParserType to complex here for use of advanced syntax
-    ->withNodeParserType(NodeParser::COMPLEX->value)
+    ->withRuleSetParserType(RuleSetParser::ARRAY->value)
+    ->withNodeParserType(NodeParser::ARRAY->value)
     // set faker to true here to use the default faker instance
     ->withFaker(true)
     ->build();
 
 $anonymizer->registerRuleSet(
-    'person-data',
-    [
-        'order.person.first_name[#firstName]',
-        'order.person.last_name[#lastName]',
+    name: 'order',
+    definitions: [
+        [
+            'name' => 'order',
+            'children' => [
+                [
+                    'name' => 'person',
+                    'children' => [
+                        [
+                            'name' => 'first_name',
+                            'value_type' => 'firstName',
+                        ],
+                        [
+                            'name' => 'last_name',
+                            'value_type' => 'lastName',
+                        ],
+                    ],
+                ],
+            ],
+        ],
     ],
 );
 
@@ -32,7 +49,10 @@ $data = [
     ],
 ];
 
-$anonymizedData = $anonymizer->run('person-data', $data);
+$anonymizedData = $anonymizer->run(
+    ruleSetName: 'order',
+    data: $data,
+);
 
 echo PHP_EOL . 'Original data:' . PHP_EOL;
 print_r($data);
