@@ -7,22 +7,41 @@ require_once __DIR__ . '/../vendor/autoload.php';
 use Faker\Factory;
 use PhpAnonymizer\Anonymizer\AnonymizerBuilder;
 use PhpAnonymizer\Anonymizer\Enum\NodeParser;
+use PhpAnonymizer\Anonymizer\Enum\RuleSetParser;
 
 $faker = Factory::create('de_DE');
 
 $anonymizer = (new AnonymizerBuilder())
     ->withDefaults()
-    // set nodeParserType to complex here for use of advanced syntax
-    ->withNodeParserType(NodeParser::COMPLEX->value)
-    // set faker to true here to use the default faker instance
+    ->withRuleSetParserType(RuleSetParser::ARRAY->value)
+    ->withNodeParserType(NodeParser::ARRAY->value)
+    // pass custom Faker instance here
     ->withCustomFaker($faker)
     ->build();
 
 $anonymizer->registerRuleSet(
-    'person-data',
-    [
-        'order.person.first_name[#firstName]',
-        'order.person.last_name[#lastName]',
+    name: 'order',
+    definitions: [
+        'nodes' => [
+            [
+                'name' => 'order',
+                'children' => [
+                    [
+                        'name' => 'person',
+                        'children' => [
+                            [
+                                'name' => 'first_name',
+                                'value_type' => 'firstName',
+                            ],
+                            [
+                                'name' => 'last_name',
+                                'value_type' => 'lastName',
+                            ],
+                        ],
+                    ],
+                ],
+            ],
+        ],
     ],
 );
 
@@ -35,7 +54,10 @@ $data = [
     ],
 ];
 
-$anonymizedData = $anonymizer->run('person-data', $data);
+$anonymizedData = $anonymizer->run(
+    ruleSetName: 'order',
+    data: $data,
+);
 
 echo PHP_EOL . 'Original data:' . PHP_EOL;
 print_r($data);
