@@ -5,16 +5,34 @@ declare(strict_types=1);
 require_once __DIR__ . '/../vendor/autoload.php';
 
 use PhpAnonymizer\Anonymizer\AnonymizerBuilder;
+use PhpAnonymizer\Anonymizer\Enum\NodeParser;
+use PhpAnonymizer\Anonymizer\Enum\RuleSetParser;
 
 $anonymizer = (new AnonymizerBuilder())
     ->withDefaults()
+    ->withRuleSetParserType(RuleSetParser::ARRAY->value)
+    ->withNodeParserType(NodeParser::ARRAY->value)
     ->build();
 
 $anonymizer->registerRuleSet(
-    'person-data',
-    [
-        'order.person.first_name',
-        'order.person.last_name',
+    name: 'order',
+    definitions: [
+        [
+            'name' => 'order',
+            'children' => [
+                [
+                    'name' => 'person',
+                    'children' => [
+                        [
+                            'name' => 'first_name',
+                        ],
+                        [
+                            'name' => 'last_name',
+                        ],
+                    ],
+                ],
+            ],
+        ],
     ],
 );
 
@@ -27,16 +45,25 @@ $data = [
     ],
 ];
 
-$anonymizedData = $anonymizer->run('person-data', $data, 'noop');
+$anonymizedData = $anonymizer->run(
+    ruleSetName: 'order',
+    data: $data,
+    // pass encoder to use here
+    encoding: 'noop',
+);
 
 echo PHP_EOL . 'Original data:' . PHP_EOL;
 print_r($data);
+echo PHP_EOL;
 
 echo PHP_EOL . 'Anonymized data:' . PHP_EOL;
 print_r($anonymizedData);
+echo PHP_EOL;
 
 echo PHP_EOL . 'Check $data == $anonymizedData' . PHP_EOL;
 var_dump($data == $anonymizedData);
+echo PHP_EOL;
 
 echo PHP_EOL . 'Check $data === $anonymizedData' . PHP_EOL;
 var_dump($data === $anonymizedData);
+echo PHP_EOL;
