@@ -3,46 +3,66 @@
 declare(strict_types=1);
 
 require_once __DIR__ . '/../vendor/autoload.php';
-require_once __DIR__ . '/includes/person_class.php';
-require_once __DIR__ . '/includes/order_class.php';
 
 use PhpAnonymizer\Anonymizer\AnonymizerBuilder;
 use PhpAnonymizer\Anonymizer\Enum\DataAccess;
 use PhpAnonymizer\Anonymizer\Enum\NodeParser;
+use PhpAnonymizer\Anonymizer\Enum\RuleSetParser;
+use PhpAnonymizer\Anonymizer\Examples\Order;
+use PhpAnonymizer\Anonymizer\Examples\Person;
 
 $anonymizer = (new AnonymizerBuilder())
     ->withDefaults()
-    ->withNodeParserType(NodeParser::COMPLEX->value)
+    ->withRuleSetParserType(RuleSetParser::ARRAY->value)
+    ->withNodeParserType(NodeParser::ARRAY->value)
     ->build();
 
 $anonymizer->registerRuleSet(
-    'person-data',
-    [
-        'person.firstName',
-        'person.lastName',
+    name: 'order',
+    definitions: [
+        [
+            'name' => 'person',
+            'children' => [
+                [
+                    'name' => 'firstName',
+                ],
+                [
+                    'name' => 'lastName',
+                ],
+            ],
+        ],
     ],
-    DataAccess::AUTODETECT->value,
+    defaultDataAccess: DataAccess::AUTODETECT->value,
 );
 
 $person = new Person(
-    'John',
-    'Doe'
+    firstName: 'John',
+    lastName: 'Doe',
 );
 
 $data = new Order(
-    $person,
+    person: $person,
 );
 
-$anonymizedData = $anonymizer->run('person-data', $data, 'clone');
+$anonymizedData = $anonymizer->run(
+    ruleSetName: 'order',
+    data: $data,
+    // pass encoder to use here
+    encoding: 'clone',
+);
 
 echo PHP_EOL . 'Original data:' . PHP_EOL;
 print_r($data);
+echo PHP_EOL;
 
 echo PHP_EOL . 'Anonymized data:' . PHP_EOL;
 print_r($anonymizedData);
+echo PHP_EOL;
 
 echo PHP_EOL . 'Check $data == $anonymizedData' . PHP_EOL;
 var_dump($data == $anonymizedData);
+echo PHP_EOL;
 
 echo PHP_EOL . 'Check $data === $anonymizedData' . PHP_EOL;
 var_dump($data === $anonymizedData);
+echo PHP_EOL;
