@@ -4,6 +4,6 @@ declare(strict_types=1);
 
 namespace PhpAnonymizer\Anonymizer\Exception;
 
-class CannotResolveValueException extends BaseAnonymizerException
+final class CannotResolveValueException extends BaseAnonymizerException
 {
 }
