@@ -4,6 +4,6 @@ declare(strict_types=1);
 
 namespace PhpAnonymizer\Anonymizer\Exception;
 
-class UnknownRuleSetException extends BaseAnonymizerException
+final class UnknownRuleSetException extends BaseAnonymizerException
 {
 }
