@@ -11,16 +11,18 @@ use PhpAnonymizer\Anonymizer\DataGeneration\Provider\DataGenerationProviderInter
 use PhpAnonymizer\Anonymizer\Enum\DataAccess;
 use PhpAnonymizer\Anonymizer\Enum\NodeType;
 use PhpAnonymizer\Anonymizer\Exception\DataEncodingException;
+use PhpAnonymizer\Anonymizer\Exception\InvalidObjectTypeException;
 use function sprintf;
 
-class ProcessingUnit
+final class ProcessingUnit
 {
-    private ?string $overrideDataAccess;
+    private ?string $overrideDataAccess = null;
 
     public function __construct(
         private readonly DataGenerationProviderInterface $dataGenerationProvider,
         private readonly DataAccessProviderInterface $dataAccessProvider,
         private readonly ?DataEncodingProviderInterface $dataEncodingProvider,
+        private readonly RuleSetProviderInterface $ruleSetProvider,
         private readonly RuleSet $ruleSet,
         private readonly mixed $data,
         private readonly TempStorage $tempStorage = new TempStorage(),
@@ -49,19 +51,36 @@ class ProcessingUnit
     /**
      * @param string[] $path
      */
-    private function processNode(array $path, mixed &$data, Node $rule): mixed
+    private function processNode(array $path, mixed &$data, Node $node): mixed
     {
-        $path[] = $rule->name;
+        $path[] = $node->name;
 
-        $dataAccess = $this->getDataAccess($rule->dataAccess);
-        if (!$dataAccess->hasChild($path, $data, $rule->name)) {
+        $dataAccess = $this->getDataAccess($node->dataAccess);
+        if (!$dataAccess->hasChild($path, $data, $node->name)) {
             return $data;
         }
 
-        $value = $dataAccess->getChild($path, $data, $rule->name);
-        $value = $this->processValue($path, $rule, $value);
+        $value = $dataAccess->getChild($path, $data, $node->name);
 
-        $dataAccess->setChildValue($path, $data, $rule->name, $value);
+        if (!$node->hasFilterRule() || ($node->filterValue === $dataAccess->getChild($path, $data, $node->filterField))) {
+            if ($node->containsNestedData()) {
+                $nestedUnit = new ProcessingUnit(
+                    $this->dataGenerationProvider,
+                    $this->dataAccessProvider,
+                    $this->dataEncodingProvider,
+                    $this->ruleSetProvider,
+                    $this->ruleSetProvider->getRuleSet($node->nestedRule),
+                    $value,
+                    new TempStorage(),
+                );
+
+                $value = $nestedUnit->process($node->nestedType);
+            } else {
+                $value = $this->processValue($path, $node, $value);
+            }
+
+            $dataAccess->setChildValue($path, $data, $node->name, $value);
+        }
 
         return $data;
     }
@@ -72,6 +91,10 @@ class ProcessingUnit
     private function processValue(array $path, Node $rule, mixed $value): mixed
     {
         if ($rule->isArray) {
+            if (!is_array($value)) {
+                throw InvalidObjectTypeException::notAnArray($path);
+            }
+
             return $this->processListValue($path, $rule, $value);
         }
 
