@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace PhpAnonymizer\Anonymizer\Model;
 
-readonly class RuleSet
+final readonly class RuleSet
 {
     public function __construct(
         public Tree $tree,
