@@ -8,25 +8,33 @@ use PhpAnonymizer\Anonymizer\DataAccess\Provider\DataAccessProviderInterface;
 use PhpAnonymizer\Anonymizer\DataEncoding\Provider\DataEncodingProviderInterface;
 use PhpAnonymizer\Anonymizer\DataGeneration\Provider\DataGenerationProviderInterface;
 use PhpAnonymizer\Anonymizer\Model\ProcessingUnit;
-use PhpAnonymizer\Anonymizer\Model\RuleSet;
+use PhpAnonymizer\Anonymizer\Model\RuleSetProvider;
+use PhpAnonymizer\Anonymizer\Model\RuleSetProviderInterface;
 
-readonly class DefaultDataProcessor implements DataProcessorInterface
+final readonly class DefaultDataProcessor implements DataProcessorInterface
 {
     public function __construct(
         private DataAccessProviderInterface $dataAccessProvider,
         private DataGenerationProviderInterface $dataGenerationProvider,
-        private DataEncodingProviderInterface $dataTransformationProvider,
+        private DataEncodingProviderInterface $dataEncodingProvider,
+        private RuleSetProviderInterface $ruleSetProvider = new RuleSetProvider(),
     ) {
     }
 
-    public function process(mixed $data, RuleSet $ruleSet, ?string $encoding = null): mixed
+    public function process(mixed $data, string $ruleSetName, ?string $encoding = null): mixed
     {
         return (new ProcessingUnit(
             dataGenerationProvider: $this->dataGenerationProvider,
             dataAccessProvider: $this->dataAccessProvider,
-            dataEncodingProvider: $this->dataTransformationProvider,
-            ruleSet: $ruleSet,
+            dataEncodingProvider: $this->dataEncodingProvider,
+            ruleSetProvider: $this->ruleSetProvider,
+            ruleSet: $this->ruleSetProvider->getRuleSet($ruleSetName),
             data: $data,
         ))->process($encoding);
+    }
+
+    public function getRuleSetProvider(): RuleSetProviderInterface
+    {
+        return $this->ruleSetProvider;
     }
 }
