@@ -6,13 +6,15 @@ namespace PhpAnonymizer\Anonymizer\Test\Unit\DataAccess;
 
 use PhpAnonymizer\Anonymizer\DataAccess\PropertyDataAccess;
 use PhpAnonymizer\Anonymizer\Exception\FieldDoesNotExistException;
+use PhpAnonymizer\Anonymizer\Exception\FieldIsNotInitializedException;
 use PhpAnonymizer\Anonymizer\Exception\InvalidObjectTypeException;
+use PhpAnonymizer\Anonymizer\Test\Helper\Model\Barfoo;
 use PhpAnonymizer\Anonymizer\Test\Helper\Model\Foobar;
 use PhpAnonymizer\Anonymizer\Test\Helper\Model\ReadonlyFoobar;
 use PHPUnit\Framework\TestCase;
 use stdClass;
 
-class PropertyDataAccessTest extends TestCase
+final class PropertyDataAccessTest extends TestCase
 {
     public function testCanCheckIfChildPropertyExists(): void
     {
@@ -21,9 +23,17 @@ class PropertyDataAccessTest extends TestCase
         $data = new stdClass();
         $data->foo = 'bar';
 
-        $this->assertTrue($access->hasChild(['test'], $data, 'foo'));
-        $this->assertFalse($access->hasChild(['test'], $data, 'bar'));
-        $this->assertFalse($access->hasChild(['test'], $data, 'FOO'));
+        self::assertTrue($access->hasChild(['test'], $data, 'foo'));
+        self::assertFalse($access->hasChild(['test'], $data, 'bar'));
+        self::assertFalse($access->hasChild(['test'], $data, 'FOO'));
+    }
+
+    public function testCanCheckIfUnitializedChildPropertyDoesNotExist(): void
+    {
+        $access = new PropertyDataAccess();
+
+        $data = new Barfoo();
+        self::assertFalse($access->hasChild(['test'], $data, 'foo'));
     }
 
     public function testWillFailOnCheckForChildPropertyOfNonObject(): void
@@ -45,7 +55,7 @@ class PropertyDataAccessTest extends TestCase
         $data = new stdClass();
         $data->foo = 'bar';
 
-        $this->assertSame('bar', $access->getChild(['test'], $data, 'foo'));
+        self::assertSame('bar', $access->getChild(['test'], $data, 'foo'));
     }
 
     public function testWillFailOnRetrieveValueOfChildPropertyOnNonObject(): void
@@ -71,6 +81,16 @@ class PropertyDataAccessTest extends TestCase
         $access->getChild(['test'], $data, 'bar');
     }
 
+    public function testWillFailOnRetrieveValueOfUninitializedChildProperty(): void
+    {
+        $access = new PropertyDataAccess();
+
+        $data = new Barfoo();
+
+        $this->expectException(FieldIsNotInitializedException::class);
+        $access->getChild(['test'], $data, 'foo');
+    }
+
     public function testWillFailOnRetrieveValueOfPrivateChildProperty(): void
     {
         $access = new PropertyDataAccess();
@@ -94,8 +114,8 @@ class PropertyDataAccessTest extends TestCase
 
         $access->setChildValue(['test'], $data, 'bar', 'baz');
 
-        $this->assertObjectHasProperty('bar', $data);
-        $this->assertSame('baz', $data->bar);
+        self::assertObjectHasProperty('bar', $data);
+        self::assertSame('baz', $data->bar);
     }
 
     public function testWillFailOnSetValueOfChildPropertyOfInvalidType(): void
@@ -142,13 +162,13 @@ class PropertyDataAccessTest extends TestCase
     {
         $access = new PropertyDataAccess();
 
-        $this->assertTrue($access->supports(new stdClass()));
-        $this->assertTrue($access->supports(new Foobar(
+        self::assertTrue($access->supports(new stdClass()));
+        self::assertTrue($access->supports(new Foobar(
             foo: 'foo',
             bar: 'bar',
             baz: 'baz',
         )));
-        $this->assertFalse($access->supports([]));
-        $this->assertFalse($access->supports('foobar'));
+        self::assertFalse($access->supports([]));
+        self::assertFalse($access->supports('foobar'));
     }
 }
