@@ -16,30 +16,30 @@ use PhpAnonymizer\Anonymizer\Exception\DataAccessExistsException;
 use PhpAnonymizer\Anonymizer\Exception\UnknownDataAccessException;
 use PHPUnit\Framework\TestCase;
 
-class DefaultDataAccessProviderTest extends TestCase
+final class DefaultDataAccessProviderTest extends TestCase
 {
     public function testCanVerifySupportOfDataAccessTypes(): void
     {
         $provider = new DefaultDataAccessProvider();
 
-        $this->assertTrue($provider->supports(DataAccess::ARRAY->value));
-        $this->assertTrue($provider->supports(DataAccess::AUTODETECT->value));
-        $this->assertTrue($provider->supports(DataAccess::PROPERTY->value));
-        $this->assertTrue($provider->supports(DataAccess::SETTER->value));
-        $this->assertTrue($provider->supports(DataAccess::REFLECTION->value));
-        $this->assertFalse($provider->supports(DataAccess::DEFAULT->value));
-        $this->assertFalse($provider->supports('foobar'));
+        self::assertTrue($provider->supports(DataAccess::ARRAY->value));
+        self::assertTrue($provider->supports(DataAccess::AUTODETECT->value));
+        self::assertTrue($provider->supports(DataAccess::PROPERTY->value));
+        self::assertTrue($provider->supports(DataAccess::SETTER->value));
+        self::assertTrue($provider->supports(DataAccess::REFLECTION->value));
+        self::assertFalse($provider->supports(DataAccess::DEFAULT->value));
+        self::assertFalse($provider->supports('foobar'));
     }
 
     public function testCanProvideAllDefaultDataAccesses(): void
     {
         $provider = new DefaultDataAccessProvider();
 
-        $this->assertInstanceOf(ArrayDataAccess::class, $provider->provideDataAccess(DataAccess::ARRAY->value));
-        $this->assertInstanceOf(AutoDetectDataAccess::class, $provider->provideDataAccess(DataAccess::AUTODETECT->value));
-        $this->assertInstanceOf(PropertyDataAccess::class, $provider->provideDataAccess(DataAccess::PROPERTY->value));
-        $this->assertInstanceOf(ReflectionDataAccess::class, $provider->provideDataAccess(DataAccess::REFLECTION->value));
-        $this->assertInstanceOf(SetterDataAccess::class, $provider->provideDataAccess(DataAccess::SETTER->value));
+        self::assertInstanceOf(ArrayDataAccess::class, $provider->provideDataAccess(DataAccess::ARRAY->value));
+        self::assertInstanceOf(AutoDetectDataAccess::class, $provider->provideDataAccess(DataAccess::AUTODETECT->value));
+        self::assertInstanceOf(PropertyDataAccess::class, $provider->provideDataAccess(DataAccess::PROPERTY->value));
+        self::assertInstanceOf(ReflectionDataAccess::class, $provider->provideDataAccess(DataAccess::REFLECTION->value));
+        self::assertInstanceOf(SetterDataAccess::class, $provider->provideDataAccess(DataAccess::SETTER->value));
     }
 
     public function testWillFailOnProvidingDefaultDataAccess(): void
@@ -65,10 +65,10 @@ class DefaultDataAccessProviderTest extends TestCase
         $dataAccess = $this->createMock(DataAccessInterface::class);
         $provider->registerCustomDataAccess('foobar', $dataAccess);
 
-        $this->assertTrue($provider->supports('foobar'));
+        self::assertTrue($provider->supports('foobar'));
 
         $resolvedDataAccess = $provider->provideDataAccess('foobar');
-        $this->assertSame($dataAccess, $resolvedDataAccess);
+        self::assertSame($dataAccess, $resolvedDataAccess);
 
         $this->expectException(DataAccessExistsException::class);
         $provider->registerCustomDataAccess('foobar', $dataAccess);
