@@ -6,13 +6,15 @@ namespace PhpAnonymizer\Anonymizer\Test\Unit\DataAccess;
 
 use PhpAnonymizer\Anonymizer\DataAccess\ReflectionDataAccess;
 use PhpAnonymizer\Anonymizer\Exception\FieldDoesNotExistException;
+use PhpAnonymizer\Anonymizer\Exception\FieldIsNotInitializedException;
 use PhpAnonymizer\Anonymizer\Exception\InvalidObjectTypeException;
+use PhpAnonymizer\Anonymizer\Test\Helper\Model\Barfoo;
 use PhpAnonymizer\Anonymizer\Test\Helper\Model\Foobar;
 use PhpAnonymizer\Anonymizer\Test\Helper\Model\ReadonlyFoobar;
 use PHPUnit\Framework\TestCase;
 use stdClass;
 
-class ReflectionDataAccessTest extends TestCase
+final class ReflectionDataAccessTest extends TestCase
 {
     public function testCanCheckIfChildPropertyExists(): void
     {
@@ -22,9 +24,17 @@ class ReflectionDataAccessTest extends TestCase
             public string $foo = 'bar';
         });
 
-        $this->assertTrue($access->hasChild(['test'], $data, 'foo'));
-        $this->assertFalse($access->hasChild(['test'], $data, 'bar'));
-        $this->assertFalse($access->hasChild(['test'], $data, 'FOO'));
+        self::assertTrue($access->hasChild(['test'], $data, 'foo'));
+        self::assertFalse($access->hasChild(['test'], $data, 'bar'));
+        self::assertFalse($access->hasChild(['test'], $data, 'FOO'));
+    }
+
+    public function testCanCheckIfUnitializedChildPropertyDoesNotExist(): void
+    {
+        $access = new ReflectionDataAccess();
+
+        $data = new Barfoo();
+        self::assertFalse($access->hasChild(['test'], $data, 'foo'));
     }
 
     public function testWillFailOnCheckForChildPropertyOfNonObject(): void
@@ -47,7 +57,7 @@ class ReflectionDataAccessTest extends TestCase
             public string $foo = 'bar';
         });
 
-        $this->assertSame('bar', $access->getChild(['test'], $data, 'foo'));
+        self::assertSame('bar', $access->getChild(['test'], $data, 'foo'));
     }
 
     public function testWillFailOnRetrieveValueOfChildPropertyOnNonObject(): void
@@ -74,6 +84,16 @@ class ReflectionDataAccessTest extends TestCase
         $access->getChild(['test'], $data, 'bar');
     }
 
+    public function testWillFailOnRetrieveValueOfUninitializedChildProperty(): void
+    {
+        $access = new ReflectionDataAccess();
+
+        $data = new Barfoo();
+
+        $this->expectException(FieldIsNotInitializedException::class);
+        $access->getChild(['test'], $data, 'foo');
+    }
+
     public function testCanRetrieveValueOfPrivateChildProperty(): void
     {
         $access = new ReflectionDataAccess();
@@ -85,7 +105,7 @@ class ReflectionDataAccessTest extends TestCase
         );
 
         $value = $access->getChild(['test'], $data, 'foo');
-        $this->assertSame('foo', $value);
+        self::assertSame('foo', $value);
     }
 
     public function testCanSetValueOfChildProperty(): void
@@ -98,8 +118,8 @@ class ReflectionDataAccessTest extends TestCase
 
         $access->setChildValue(['test'], $data, 'bar', 'baz');
 
-        $this->assertObjectHasProperty('bar', $data);
-        $this->assertSame('baz', $data->bar);
+        self::assertObjectHasProperty('bar', $data);
+        self::assertSame('baz', $data->bar);
     }
 
     public function testWillFailOnSetValueOfChildPropertyOfInvalidType(): void
@@ -144,13 +164,13 @@ class ReflectionDataAccessTest extends TestCase
     {
         $access = new ReflectionDataAccess();
 
-        $this->assertTrue($access->supports(new stdClass()));
-        $this->assertTrue($access->supports(new Foobar(
+        self::assertTrue($access->supports(new stdClass()));
+        self::assertTrue($access->supports(new Foobar(
             foo: 'foo',
             bar: 'bar',
             baz: 'baz',
         )));
-        $this->assertFalse($access->supports([]));
-        $this->assertFalse($access->supports('foobar'));
+        self::assertFalse($access->supports([]));
+        self::assertFalse($access->supports('foobar'));
     }
 }
