@@ -6,12 +6,14 @@ namespace PhpAnonymizer\Anonymizer\Test\Unit\DataAccess;
 
 use PhpAnonymizer\Anonymizer\DataAccess\SetterDataAccess;
 use PhpAnonymizer\Anonymizer\Exception\FieldDoesNotExistException;
+use PhpAnonymizer\Anonymizer\Exception\FieldIsNotInitializedException;
 use PhpAnonymizer\Anonymizer\Exception\InvalidObjectTypeException;
+use PhpAnonymizer\Anonymizer\Test\Helper\Model\Barfoo;
 use PhpAnonymizer\Anonymizer\Test\Helper\Model\Foobar;
 use PHPUnit\Framework\TestCase;
 use stdClass;
 
-class SetterDataAccessTest extends TestCase
+final class SetterDataAccessTest extends TestCase
 {
     public function testCanCheckIfChildPropertyExists(): void
     {
@@ -23,9 +25,17 @@ class SetterDataAccessTest extends TestCase
             baz: 'baz',
         );
 
-        $this->assertTrue($access->hasChild(['test'], $data, 'baz'));
-        $this->assertFalse($access->hasChild(['test'], $data, 'foo'));
-        $this->assertFalse($access->hasChild(['test'], $data, 'bar'));
+        self::assertTrue($access->hasChild(['test'], $data, 'baz'));
+        self::assertFalse($access->hasChild(['test'], $data, 'foo'));
+        self::assertFalse($access->hasChild(['test'], $data, 'bar'));
+    }
+
+    public function testCanCheckIfUnitializedChildPropertyDoesNotExist(): void
+    {
+        $access = new SetterDataAccess();
+
+        $data = new Barfoo();
+        self::assertFalse($access->hasChild(['test'], $data, 'foo'));
     }
 
     public function testWillFailOnCheckForChildPropertyOfNonObject(): void
@@ -50,7 +60,7 @@ class SetterDataAccessTest extends TestCase
             baz: 'baz',
         );
 
-        $this->assertSame('baz', $access->getChild(['test'], $data, 'baz'));
+        self::assertSame('baz', $access->getChild(['test'], $data, 'baz'));
     }
 
     public function testWillFailOnRetrieveValueOfChildPropertyOnNonObject(): void
@@ -79,6 +89,16 @@ class SetterDataAccessTest extends TestCase
         $access->getChild(['test'], $data, 'foo2');
     }
 
+    public function testWillFailOnRetrieveValueOfUninitializedChildProperty(): void
+    {
+        $access = new SetterDataAccess();
+
+        $data = new Barfoo();
+
+        $this->expectException(FieldIsNotInitializedException::class);
+        $access->getChild(['test'], $data, 'foo');
+    }
+
     public function testWillFailOnRetrieveValueOfWriteonlyChildProperty(): void
     {
         $access = new SetterDataAccess();
@@ -104,7 +124,7 @@ class SetterDataAccessTest extends TestCase
         );
 
         $access->setChildValue(['test'], $data, 'baz', 'new baz');
-        $this->assertSame('new baz', $data->getBaz());
+        self::assertSame('new baz', $data->getBaz());
     }
 
     public function testWillFailOnSetValueOfChildPropertyOfInvalidType(): void
@@ -151,13 +171,13 @@ class SetterDataAccessTest extends TestCase
     {
         $access = new SetterDataAccess();
 
-        $this->assertTrue($access->supports(new stdClass()));
-        $this->assertTrue($access->supports(new Foobar(
+        self::assertTrue($access->supports(new stdClass()));
+        self::assertTrue($access->supports(new Foobar(
             foo: 'foo',
             bar: 'bar',
             baz: 'baz',
         )));
-        $this->assertFalse($access->supports([]));
-        $this->assertFalse($access->supports('foobar'));
+        self::assertFalse($access->supports([]));
+        self::assertFalse($access->supports('foobar'));
     }
 }
