@@ -18,7 +18,7 @@ use RuntimeException;
 use stdClass;
 use Symfony\Component\Serializer\Normalizer\ObjectNormalizer;
 
-class DefaultDataEncodingProviderTest extends TestCase
+final class DefaultDataEncodingProviderTest extends TestCase
 {
     public function testWillFailOnInitializationWhenSymfonyPackageIsMissingAndNormalizerIsGiven(): void
     {
@@ -81,7 +81,7 @@ class DefaultDataEncodingProviderTest extends TestCase
         $provider->setDenormalizer($objectNormalizer);
 
         $encoder = $provider->provideEncoder(DataEncoder::SYMFONY->value);
-        $this->assertInstanceOf(SymfonyEncoder::class, $encoder);
+        self::assertInstanceOf(SymfonyEncoder::class, $encoder);
     }
 
     public function testWillFailOnNormalizerSetterInjectionWhenSymfonyPackageIsMissing(): void
@@ -97,7 +97,6 @@ class DefaultDataEncodingProviderTest extends TestCase
 
         $this->expectException(MissingPlatformRequirementsException::class);
 
-        /** @phpstan-ignore-next-line  */
         $provider->setNormalizer(new stdClass());
     }
 
@@ -107,7 +106,6 @@ class DefaultDataEncodingProviderTest extends TestCase
 
         $this->expectException(InvalidArgumentException::class);
 
-        /** @phpstan-ignore-next-line  */
         $provider->setNormalizer(new stdClass());
     }
 
@@ -124,7 +122,6 @@ class DefaultDataEncodingProviderTest extends TestCase
 
         $this->expectException(MissingPlatformRequirementsException::class);
 
-        /** @phpstan-ignore-next-line  */
         $provider->setDenormalizer(new stdClass());
     }
 
@@ -134,7 +131,6 @@ class DefaultDataEncodingProviderTest extends TestCase
 
         $this->expectException(InvalidArgumentException::class);
 
-        /** @phpstan-ignore-next-line  */
         $provider->setDenormalizer(new stdClass());
     }
 
@@ -143,7 +139,7 @@ class DefaultDataEncodingProviderTest extends TestCase
         $provider = new DefaultDataEncodingProvider();
 
         $encoder = $provider->provideEncoder('noop');
-        $this->assertInstanceOf(NoOpEncoder::class, $encoder);
+        self::assertInstanceOf(NoOpEncoder::class, $encoder);
     }
 
     public function testWillFailOnProvideUnknownDataEncoder(): void
@@ -162,7 +158,7 @@ class DefaultDataEncodingProviderTest extends TestCase
         $provider->registerCustomDataEncoder('custom', $encoder);
 
         $resolvedEncoder = $provider->provideEncoder('custom');
-        $this->assertSame($encoder, $resolvedEncoder);
+        self::assertSame($encoder, $resolvedEncoder);
     }
 
     public function testWillFailOnRegisteringCustomEncoderOnNameConflict(): void
