@@ -9,20 +9,20 @@ use PhpAnonymizer\Anonymizer\Exception\InvalidObjectTypeException;
 use PHPUnit\Framework\TestCase;
 use stdClass;
 
-class StarMaskedStringGeneratorTest extends TestCase
+final class StarMaskedStringGeneratorTest extends TestCase
 {
     public function testCanVerifySupportOfStrings(): void
     {
         $dataGenerator = new StarMaskedStringGenerator();
 
-        $this->assertTrue($dataGenerator->supports('Hello World', null));
+        self::assertTrue($dataGenerator->supports('Hello World', null));
     }
 
     public function testCanVerifyNonSupportOfObjects(): void
     {
         $dataGenerator = new StarMaskedStringGenerator();
 
-        $this->assertFalse($dataGenerator->supports(new stdClass(), null));
+        self::assertFalse($dataGenerator->supports(new stdClass(), null));
     }
 
     public function testCanHideData(): void
@@ -32,7 +32,7 @@ class StarMaskedStringGeneratorTest extends TestCase
         $string = 'Hello World';
         $anonymizedString = $dataGenerator->generate(['test'], $string, null);
 
-        $this->assertSame('***********', $anonymizedString);
+        self::assertSame('***********', $anonymizedString);
     }
 
     public function testWillFailOnHideDataOnNonStrings(): void
