@@ -8,54 +8,46 @@ use PhpAnonymizer\Anonymizer\Enum\NodeType;
 use PhpAnonymizer\Anonymizer\Exception\InvalidArgumentException;
 use PhpAnonymizer\Anonymizer\Exception\InvalidNodeNameException;
 use PhpAnonymizer\Anonymizer\Exception\NodeDefinitionMismatchException;
-use PhpAnonymizer\Anonymizer\Model\Node;
-use PhpAnonymizer\Anonymizer\Model\Tree;
-use PhpAnonymizer\Anonymizer\Parser\Node\ComplexRegexParser;
+use PhpAnonymizer\Anonymizer\Parser\Node\ComplexRegexpParser;
 use PhpAnonymizer\Anonymizer\Parser\RuleSet\DefaultRuleSetParser;
 use PHPUnit\Framework\TestCase;
 use stdClass;
 
-class ComplexDefinitionParserTest extends TestCase
+final class ComplexDefinitionParserTest extends TestCase
 {
     public function testCanParseMultiLevelTreeDefinition(): void
     {
-        $parser = new DefaultRuleSetParser(new ComplexRegexParser());
+        $parser = new DefaultRuleSetParser(new ComplexRegexpParser());
         $tree = $parser->parseDefinition(
             [
                 'data[array].[]address[property].name[setter]',
             ],
         );
 
-        $this->assertInstanceOf(Tree::class, $tree);
-
         $dataLevel = $tree->getChildNode('data');
-        $this->assertInstanceOf(Node::class, $dataLevel);
-        $this->assertSame('data', $dataLevel->name);
-        $this->assertSame('array', $dataLevel->dataAccess);
-        $this->assertSame(NodeType::NODE, $dataLevel->nodeType);
-        $this->assertFalse($dataLevel->isArray);
+        self::assertSame('data', $dataLevel->name);
+        self::assertSame('array', $dataLevel->dataAccess);
+        self::assertSame(NodeType::NODE, $dataLevel->nodeType);
+        self::assertFalse($dataLevel->isArray);
 
         $addressLevel = $dataLevel->getChildNode('address');
-        $this->assertInstanceOf(Node::class, $addressLevel);
-        $this->assertSame('address', $addressLevel->name);
-        $this->assertSame('property', $addressLevel->dataAccess);
-        $this->assertSame(NodeType::NODE, $addressLevel->nodeType);
-        $this->assertTrue($addressLevel->isArray);
+        self::assertSame('address', $addressLevel->name);
+        self::assertSame('property', $addressLevel->dataAccess);
+        self::assertSame(NodeType::NODE, $addressLevel->nodeType);
+        self::assertTrue($addressLevel->isArray);
 
         $nameLevel = $addressLevel->getChildNode('name');
-        $this->assertInstanceOf(Node::class, $nameLevel);
-        $this->assertSame('name', $nameLevel->name);
-        $this->assertSame('setter', $nameLevel->dataAccess);
-        $this->assertSame(NodeType::LEAF, $nameLevel->nodeType);
-        $this->assertFalse($nameLevel->isArray);
+        self::assertSame('name', $nameLevel->name);
+        self::assertSame('setter', $nameLevel->dataAccess);
+        self::assertSame(NodeType::LEAF, $nameLevel->nodeType);
+        self::assertFalse($nameLevel->isArray);
     }
 
     public function testWillFailOnParseDefinitionWithInvalidDefinitionType(): void
     {
-        $parser = new DefaultRuleSetParser(new ComplexRegexParser());
+        $parser = new DefaultRuleSetParser(new ComplexRegexpParser());
         $this->expectException(InvalidArgumentException::class);
         $parser->parseDefinition(
-            /** @phpstan-ignore-next-line  */
             [
                 new stdClass(),
             ],
@@ -64,7 +56,7 @@ class ComplexDefinitionParserTest extends TestCase
 
     public function testWillFailOnParseDefinitionWithInvalidNodeName(): void
     {
-        $parser = new DefaultRuleSetParser(new ComplexRegexParser());
+        $parser = new DefaultRuleSetParser(new ComplexRegexpParser());
         $this->expectException(InvalidNodeNameException::class);
         $parser->parseDefinition(
             [
@@ -75,7 +67,7 @@ class ComplexDefinitionParserTest extends TestCase
 
     public function testWillFailOnParseDefinitionWithPrematureRuleEnding(): void
     {
-        $parser = new DefaultRuleSetParser(new ComplexRegexParser());
+        $parser = new DefaultRuleSetParser(new ComplexRegexpParser());
         $this->expectException(NodeDefinitionMismatchException::class);
         $parser->parseDefinition(
             [
@@ -87,7 +79,7 @@ class ComplexDefinitionParserTest extends TestCase
 
     public function testWillFailOnParseDefinitionWithPrematureRuleEndingScenario2(): void
     {
-        $parser = new DefaultRuleSetParser(new ComplexRegexParser());
+        $parser = new DefaultRuleSetParser(new ComplexRegexpParser());
         $this->expectException(NodeDefinitionMismatchException::class);
         $parser->parseDefinition(
             [
@@ -99,7 +91,7 @@ class ComplexDefinitionParserTest extends TestCase
 
     public function testWillFailOnParseDefinitionWithDefinitionMismatch(): void
     {
-        $parser = new DefaultRuleSetParser(new ComplexRegexParser());
+        $parser = new DefaultRuleSetParser(new ComplexRegexpParser());
         $this->expectException(NodeDefinitionMismatchException::class);
         $parser->parseDefinition(
             [
